mx_mysql: Fix serious bug when fetching empty strings (length=0)
MySQL will not mark empty strings truncated because the
length of an empty string is equal to 0 and it occupies 0 bytes
in MySQL.

So the error flag will never be set for zero length strings.

Since we need zero length strings to be \0 terminated we ignore
error conditions for string types.