mxqd: fail with diagnostic if unable to write to log directory
Failures of cronolog are not detected by the current code, because
cronologs own stderr goes to /dev/null.

While there are many reasons, why cronolog mail fail, not being able
to write to the log directory altogether is a very prominent one
which is easy to detect beforehand. This may become even more
relevant when we start to support mxqd being run as non-root.