mxqd: implement time contraint
This allows mxqd to be started with "-t <minutes>" to have a job time
contraint for the server. jobs which are submitten with a longer maximum
runtime are not started on this server.

This is an emergency implementation and might need some
cleanup/rethinking in the future. But we need the feature now.