mxqd_control: account for orphaned groups
The count of running jobs and threads of a group were
not deducted from the counters the server keeps for the user
and the cluster.

This lead to wrong scheduling, as users were thought to have more
running jobs than they really had.

Add code to maintain the global counters.